+++ lib/framework/WiFiSettingsService.h
@@ -38,7 +38,7 @@
 #define WIFI_SETTINGS_SERVICE_PATH "/rest/wifiSettings"
 
 #define WIFI_RECONNECTION_DELAY 1000 * 30
-#define RSSI_EVENT_DELAY 100
+#define RSSI_EVENT_DELAY 200
 
 class WiFiSettings
 {